+++ R/get_selected_subset.R
@@ -18,7 +18,8 @@ get_selected_subset <- function(data, ..., var = "", value = "") {
   
   df_tmp %>% 
     dplyr::select(-BetDay, -HomeTeam, -AwayTeam, -OddsGroup, 
-                  -OddsMod, -StakeGroup, -StakeMod, -ID, -GameType) %>% 
+                  -OddsMod, -StakeGroup, -StakeMod, -ID, -GameType, 
+                  -Country, -Freebet, -Bookmaker) %>% 
     dplyr::rename(Tourn = Tournament) %>% 
     dplyr::mutate(Tourn = short_tournament_name(Tourn))